Stream Processing/requirements.txt
aiokafka==0.10.0
annotated-types==0.6.0
async-timeout==4.0.3
asyncio==3.4.3
kafka-python==2.0.2
numpy==1.26.3
packaging==23.2
pandas==2.2.0
pydantic==2.6.0
pydantic_core==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
typing_extensions==4.9.0
tzdata==2023.4
